MAINT: Move curvature correction file to correction folder.
