+++ src/daria/__init__.py
@@ -9,11 +9,11 @@ from daria.utils.conversions import *
 from daria.utils.coordinatesystem import CoordinateSystem
 from daria.utils.image import Image
 from daria.utils.subregions import *
-from daria.utils.curvaturecorrection import *
 from daria.utils.derivatives import *
 from daria.utils.stoppingcriterion import *
 from daria.utils.solvers import *
 from daria.utils.norms import *
 from daria.utils.regularization import *
 from daria.corrections.color.colorchecker import *
+from daria.corrections.curvaturecorrection import *
 #from daria.image.patches import *
